# deploy/openapi/alldata-api.yaml
openapi: 3.0.3
info:
  title: ONE-DATA-STUDIO Alldata API
  description: |
    数据治理与开发平台 API

    ## 功能概述

    - 数据集 CRUD 操作
    - 元数据管理
    - MinIO 文件存储集成
    - 数据集版本管理
    - 数据预览与查询

    ## 认证方式

    API 使用 Bearer Token 认证。请在请求头中包含:
    ```
    Authorization: Bearer your-access-token
    ```
  version: 2.0.0
  contact:
    name: ONE-DATA-STUDIO Team
    email: support@one-data.example.com

servers:
  - url: http://localhost:8080
    description: 开发环境
  - url: https://api.example.com/alldata
    description: 生产环境

tags:
  - name: Health
    description: 健康检查
  - name: Datasets
    description: 数据集管理
  - name: Metadata
    description: 元数据管理

paths:
  /api/v1/health:
    get:
      tags: [Health]
      summary: 健康检查
      description: 检查 API 服务健康状态，无需认证
      operationId: health
      responses:
        '200':
          description: 服务正常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /api/v1/datasets:
    get:
      tags: [Datasets]
      summary: 获取数据集列表
      operationId: listDatasets
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, archived]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetListResponse'
    post:
      tags: [Datasets]
      summary: 创建数据集
      operationId: createDataset
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDatasetRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDatasetResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/datasets/{dataset_id}:
    get:
      tags: [Datasets]
      summary: 获取数据集详情
      operationId: getDataset
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DatasetId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetResponse'
        '404':
          description: 数据集不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Datasets]
      summary: 更新数据集
      operationId: updateDataset
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DatasetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDatasetRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetResponse'
        '404':
          description: 数据集不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Datasets]
      summary: 删除数据集
      operationId: deleteDataset
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DatasetId'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: 数据集不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/metadata/databases:
    get:
      tags: [Metadata]
      summary: 获取数据库列表
      operationId: listDatabases
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseListResponse'

  /api/v1/metadata/databases/{database}/tables:
    get:
      tags: [Metadata]
      summary: 获取数据库表列表
      operationId: listTables
      security:
        - BearerAuth: []
      parameters:
        - name: database
          in: path
          required: true
          schema:
            type: string
          description: 数据库名称
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableListResponse'

  /api/v1/metadata/databases/{database}/tables/{table}:
    get:
      tags: [Metadata]
      summary: 获取表详情
      operationId: getTableDetail
      security:
        - BearerAuth: []
      parameters:
        - name: database
          in: path
          required: true
          schema:
            type: string
          description: 数据库名称
        - name: table
          in: path
          required: true
          schema:
            type: string
          description: 表名称
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableDetailResponse'

  /api/v1/metadata/databases/{database}/tables/{table}/columns:
    get:
      tags: [Metadata]
      summary: 获取表的列信息
      operationId: listColumns
      security:
        - BearerAuth: []
      parameters:
        - name: database
          in: path
          required: true
          schema:
            type: string
        - name: table
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColumnListResponse'

  /api/v1/query:
    post:
      tags: [Datasets]
      summary: 执行数据查询
      operationId: executeQuery
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    DatasetId:
      name: dataset_id
      in: path
      required: true
      description: 数据集 ID
      schema:
        type: string

  schemas:
    HealthResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: healthy
        service:
          type: string
          example: alldata-api
        version:
          type: string
          example: 2.0.0
        database:
          type: string
          example: connected
        minio:
          type: string
          example: connected

    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 40001
        message:
          type: string
          example: Invalid request
        details:
          type: object

    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success

    Dataset:
      type: object
      properties:
        id:
          type: integer
        dataset_id:
          type: string
        name:
          type: string
        description:
          type: string
        storage_type:
          type: string
          enum: [s3, hdfs, local]
        storage_path:
          type: string
        format:
          type: string
          enum: [csv, json, parquet, excel]
        status:
          type: string
          enum: [active, inactive, archived]
        tags:
          type: array
          items:
            type: string
        row_count:
          type: integer
        size_bytes:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        columns:
          type: array
          items:
            $ref: '#/components/schemas/DatasetColumn'

    DatasetColumn:
      type: object
      properties:
        id:
          type: integer
        column_name:
          type: string
        column_type:
          type: string
        is_nullable:
          type: boolean
        description:
          type: string
        position:
          type: integer

    DatasetListResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/Dataset'

    CreateDatasetRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '^[a-zA-Z0-9\\u4e00-\\u9fa5_\\-\\s]+$'
        description:
          type: string
          maxLength: 500
        storage_type:
          type: string
          enum: [s3, hdfs, local]
          default: s3
        storage_path:
          type: string
        format:
          type: string
          enum: [csv, json, parquet, excel]
          default: csv
        tags:
          type: array
          items:
            type: string
          maxItems: 20
        schema:
          type: object
          properties:
            columns:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  type:
                    type: string

    CreateDatasetResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            dataset_id:
              type: string

    UpdateDatasetRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, archived]
        storage_path:
          type: string

    DatasetResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/Dataset'

    Database:
      type: object
      properties:
        name:
          type: string
        tables_count:
          type: integer

    DatabaseListResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            databases:
              type: array
              items:
                $ref: '#/components/schemas/Database'

    Table:
      type: object
      properties:
        name:
          type: string
        database:
          type: string
        rows_count:
          type: integer

    TableListResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            tables:
              type: array
              items:
                $ref: '#/components/schemas/Table'

    TableDetailResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            name:
              type: string
            database:
              type: string
            columns:
              type: array
              items:
                $ref: '#/components/schemas/Column'
            relations:
              type: array
              items:
                type: object
                properties:
                  from_table:
                    type: string
                  from_column:
                    type: string
                  to_table:
                    type: string
                  to_column:
                    type: string

    Column:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        primary_key:
          type: boolean
        nullable:
          type: boolean
        description:
          type: string

    ColumnListResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            columns:
              type: array
              items:
                $ref: '#/components/schemas/Column'

    QueryRequest:
      type: object
      required:
        - dataset_id
        - sql
      properties:
        dataset_id:
          type: string
        sql:
          type: string
        limit:
          type: integer
          minimum: 1
          maximum: 10000
          default: 100

    QueryResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            columns:
              type: array
              items:
                type: string
            rows:
              type: array
              items:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: number
                    - type: boolean
                    - type: null
            row_count:
              type: integer